checkbox

+++ src/app/pages/appointments/appointments.component.ts
@@ -40,7 +40,9 @@ export class AppointmentsComponent implements OnInit {
   }
 
   deleteAppointments() {
-    this.markedAppointments.forEach(appointmentGUID => this.appointmentService.deleteAppointments(appointmentGUID));
+    if (confirm('Are you sure you want to delete this appointment?')) {
+      this.markedAppointments.forEach(appointmentGUID => this.appointmentService.deleteAppointments(appointmentGUID));
+    }
     this.appointments = this.appointmentService.getAllDoctorAppointments(this.currentUser.guid);
 
   }
